+++ src/main/java/com/github/kkieffer/jzeld/element/ZImage.java
@@ -42,7 +42,7 @@ public class ZImage extends ZRectangle {
      * @param fillColor color of the rectangle area, which can be null for transparent. Any transparent image pixels will have this color
      * @param img the image painted on this element, if null, it is not painted
      */
-    public ZImage(double x, double y, double width, double height, double rotation, boolean canSelect, boolean canResize, int borderWidth, Color borderColor, Float[] dashPattern, Color fillColor, BufferedImage img) {
+    public ZImage(double x, double y, double width, double height, double rotation, boolean canSelect, boolean canResize, float borderWidth, Color borderColor, Float[] dashPattern, Color fillColor, BufferedImage img) {
         super(x, y, width, height, rotation, canSelect, canResize, borderWidth, borderColor, dashPattern, fillColor);
         if (width <= 0 || height <= 0)
             throw new IllegalArgumentException("Width and height must be positive value");
@@ -90,7 +90,7 @@ public class ZImage extends ZRectangle {
     }
     
     @Override
-    public void setAttributes(int outlineWidth, Color outlineColor, Float[] dashPattern, Color fillColor) {
+    public void setAttributes(float outlineWidth, Color outlineColor, Float[] dashPattern, Color fillColor) {
         super.setAttributes(outlineWidth, outlineColor, dashPattern, fillColor);
         backgroundColor = null;
         hasChanges = true;
