+++ src/main/java/com/github/kkieffer/jzeld/element/ZShape.java
@@ -79,7 +79,7 @@ public class ZShape extends ZAbstractShape {
      * @param dashPattern the border dash pattern, null for solid
      * @param fillColor color of the rectangle area, which can be null for transparent (but not in combination with a zero width border)
      */
-    public ZShape(double x, double y, Shape s, double rotation, boolean canSelect, boolean canResize, int borderWidth, Color borderColor, Float[] dashPattern, Color fillColor) {
+    public ZShape(double x, double y, Shape s, double rotation, boolean canSelect, boolean canResize, float borderWidth, Color borderColor, Float[] dashPattern, Color fillColor) {
         super(x, y, s.getBounds2D().getWidth(), s.getBounds2D().getHeight(), rotation, canSelect, canResize, borderWidth, borderColor, dashPattern, fillColor);        
         this.shape = s;
     }
